no traits, back to Interface
