+++ src/Widget/TwigAware.php
@@ -7,24 +7,13 @@ namespace Bolt\Widget;
 use Twig\Environment;
 
 /**
- * Trait TwigAware - Widgets that make use of a template to render their
- * contents need to use this Trait, in order to access the current
+ * Interface TwigAware - Widgets that make use of a template to render their
+ * contents need to implement this interface, in order to access the current
  * Twig\Environment and to render a Twig template.
  */
-trait TwigAware
+interface TwigAware
 {
-    /** @var Environment */
-    protected $twig;
+    public function setTwig(Environment $twig): WidgetInterface;
 
-    public function setTwig(Environment $twig): WidgetInterface
-    {
-        $this->twig = $twig;
-
-        return $this;
-    }
-
-    public function getTwig(): Environment
-    {
-        return $this->twig;
-    }
+    public function getTwig(): Environment;
 }
